refactor imagevector package; update to nwpd v0.16.0
