+++ charts/charts.go
@@ -6,13 +6,7 @@ package charts
 
 import (
 	"embed"
-	_ "embed"
 )
-
-// Images YAML contains the contents of the images.yaml file.
-//
-//go:embed images.yaml
-var ImagesYAML string
 
 // Internal contains the internal charts
 //
