+++ go.mod
@@ -5,7 +5,7 @@ go 1.21
 require (
 	github.com/ahmetb/gen-crd-api-reference-docs v0.3.0
 	github.com/gardener/gardener v1.80.1
-	github.com/gardener/network-problem-detector v0.15.0
+	github.com/gardener/network-problem-detector v0.16.0
 	github.com/go-logr/logr v1.2.4
 	github.com/onsi/ginkgo/v2 v2.11.0
 	github.com/onsi/gomega v1.27.10
