+++ imagevector/images.yaml
@@ -9,10 +9,10 @@ images:
     name: network-problem-detector
   sourceRepository: github.com/gardener/network-problem-detector
   repository: eu.gcr.io/gardener-project/gardener/network-problem-detector
-  tag: "v0.15.0"
+  tag: "v0.16.0"
 - name: network-problem-detector-controller
   sourceRepository: github.com/gardener/network-problem-detector
   resourceId:
     name: network-problem-detector
   repository: eu.gcr.io/gardener-project/gardener/network-problem-detector
-  tag: "v0.15.0"
+  tag: "v0.16.0"
